refactor: changing conversation-list structure
