--- src/app/pages/dashboard/components/widgets/chat/conversation-list/conversation-list.component.html
@@ -1,13 +1,0 @@
-<div class="max-h-[98%] overflow-y-scroll">
-  <h1 class="text-center pt-3">Messagerie</h1>
-  <div class="border-t-2 border-t-black border-b-2 border-b-black">
-    <div *ngFor="let conversation of conversationList">
-      <app-conversation
-        username="{{ conversation.username }}"
-        lastMessage="{{ conversation.lastMessage }}"
-        messageDate="{{ conversation.messageDate }}"
-        [messagesNotSeen]="conversation.messagesNotSeen"
-      />
-    </div>
-  </div>
-</div>
